fix(nextjs): fix a tailwind + emotion (mui style lib) bug
The line 7: important: '#__next', in tailwind.config.js is necessary to fix a bug that override the tailwind classes with the original css properties of MUI5 components style.
This bug happened only in production, not in dev mode.
It was found on AppBar, that was rendered with the Login Button close to the home link, even with the class ml-auto. Debugging the css, I found out that the Tailwind's ml-auto was there, but it was disable by the priority level higher in the css generated by the Emotion library.